[resharding] Refactor `uild_state_for_split_shards` and related functions in one file
