+++ chain/chain/src/test_utils/kv_runtime.rs
@@ -12,7 +12,6 @@ use num_rational::Ratio;
 
 use near_chain_configs::{ProtocolConfig, DEFAULT_GC_NUM_EPOCHS_TO_KEEP};
 use near_chain_primitives::Error;
-use near_client_primitives::types::StateSplitApplyingStatus;
 use near_crypto::{KeyType, PublicKey, SecretKey, Signature};
 use near_pool::types::PoolIterator;
 use near_primitives::account::{AccessKey, Account};
@@ -1395,15 +1394,5 @@ impl RuntimeAdapter for KeyValueRuntime {
         _state_changes: StateChangesForSplitStates,
     ) -> Result<Vec<ApplySplitStateResult>, Error> {
         Ok(vec![])
-    }
-
-    fn build_state_for_split_shards(
-        &self,
-        _shard_uid: ShardUId,
-        _state_root: &StateRoot,
-        _next_epoch_shard_layout: &ShardLayout,
-        _state_split_status: Arc<StateSplitApplyingStatus>,
-    ) -> Result<HashMap<ShardUId, StateRoot>, Error> {
-        Ok(HashMap::new())
     }
 }
