+++ chain/chain/src/types.rs
@@ -1,5 +1,4 @@
 use std::collections::HashMap;
-use std::sync::Arc;
 
 use borsh::{BorshDeserialize, BorshSerialize};
 use chrono::DateTime;
@@ -11,7 +10,6 @@ use num_rational::Rational32;
 use crate::metrics;
 use near_chain_configs::{Genesis, ProtocolConfig};
 use near_chain_primitives::Error;
-use near_client_primitives::types::StateSplitApplyingStatus;
 use near_pool::types::PoolIterator;
 use near_primitives::challenge::ChallengesResult;
 use near_primitives::checked_feature;
@@ -437,14 +435,6 @@ pub trait RuntimeAdapter: Send + Sync {
         state_changes: StateChangesForSplitStates,
     ) -> Result<Vec<ApplySplitStateResult>, Error>;
 
-    fn build_state_for_split_shards(
-        &self,
-        shard_uid: ShardUId,
-        state_root: &StateRoot,
-        next_epoch_shard_layout: &ShardLayout,
-        state_split_status: Arc<StateSplitApplyingStatus>,
-    ) -> Result<HashMap<ShardUId, StateRoot>, Error>;
-
     /// Should be executed after accepting all the parts to set up a new state.
     fn apply_state_part(
         &self,
@@ -486,6 +476,8 @@ pub struct LatestKnown {
 
 #[cfg(test)]
 mod tests {
+    use std::sync::Arc;
+
     use chrono::Utc;
     use near_primitives::test_utils::{create_test_signer, TestBlockBuilder};
